requirements.txt
arxiv>=1.4.7
requests>=2.31.0
python-dotenv>=1.0.0
markdown>=3.4.3
pathlib>=1.0.1
python-dateutil>=2.8.2
pytz>=2022.1
python-jobspy
pandas>=1.3.6
